refactor(perl-lsp-rs): split call hierarchy inference submodule (#0000)
